utils/requirements.txt
pinecone-client==2
openai
jsonref
python-dotenv